requirements.txt
cryptography==44.0.1
Babel==2.9.1
dj-database-url==2.3.0
dj-redis-url==0.1.4
Django==4.2.22
django-db-geventpool==4.0.1
django-cors-headers==4.6.0
django-filter==24.3
django-picklefield==3.2
fyle-djangoq2==2.0.0
django-request-logging==0.7.5
djangorestframework==3.15.2
django-sendgrid-v5==1.2.0
fyle==1.0.2
fyle-accounting-mappings==2.5.5
fyle-integrations-platform-connector==2.3.2
fyle-rest-auth==1.8.3
flake8==4.0.1
gevent==24.11.1
gunicorn==23.0.0
isort==5.10.1
psycopg2-binary==2.9.10
pylint==2.7.4
python-dateutil==2.8.2
pytz==2024.2
qbosdk==1.2.0
redis==3.3.11
requests==2.32.4
sentry-sdk==2.19.2
six==1.13.0
Unidecode==1.1.2
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.9.0
pytest-mock==3.14.0
git+https://github.com/fylein/pgevents.git@main#egg=pgevents
